Add files via upload
The following request commands are available in BUKA :
• AUTH <Name> <Password>
Provide a name and password for client login, e.g.
AUTH Drizzy p455w0rd. The server should validate the provided name and password.
If the provided credentials are not valid, an error message must be returned. However
if the login is successful, the server should allow the client to make other requests.
• LIST
Return a list of available pdf files
The server retrieves the available pdf files from the text file and return the list to the
client.
• PDFRET <ID>
Return the requested pdf file to the client
The server should validate the ID and return an error message if the ID is not valid. If
the ID is valid, a confirmation message should be sent including the file size. The File
should then be read in and transmitted to the client
• LOGOUT
Log a client off.
The following responses are used in BUKA :
• 200 <Message> Successful command with <Message> providing a helpful message
from the server.
• 500 <Message> Unsuccessful command with <Message> providing a reason that the
command did not work.