Add files via upload
The following request commands are available in ZEDEM :
• BONJOUR <Name> <Password>
Provide a name and password for client login, e.g.
BONJOUR User Pass235. The server should validate the provided name and
password. If the provided credentials are not valid, an error message must
be returned. However if the login is successful, the server should allow the
client to make other requests.
• PLAYLIST
Return a list of available audio files
The server retrieves the available audio files from the List.txt text file and
return the list to the client.
• ZEDEMGET <ID>
Return a response message followed by the requested audio file to the client
The server should validate the ID and return an error message if the ID is not
valid. If the ID is valid, a confirmation message should be sent including the
file size. The File should then be read in and transmitted to the client.
• ZEDEMBYE
Log a client off.
The following responses are used in ZEDEM :
• JA <Message> Successful command with <Message> providing a helpful message from the server.
• NEE <Message> Unsuccessful command with <Message> providing a reason
that the command did not work.